# Conflicts:
#	backend/src/main/java/org/ssafy/load/application/GoodsService.java

+++ backend/src/main/java/org/ssafy/load/dao/DeliveryAreaRepository.java
@@ -3,5 +3,6 @@ package org.ssafy.load.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.ssafy.load.domain.DeliveryAreaEntity;
 
-public interface DeliveryAreaRepository extends JpaRepository<DeliveryAreaEntity, Long> {
+
+public interface DeliveryAreaRepository extends JpaRepository<DeliveryAreaEntity, Integer> {
 }
